scenarios/chatbot/simple-chatbot/pyproject.toml
[project]
name = "simple-chatbot"
version = "0.1.0"
description = "Simple chatbot by langchain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.2.1",
]
